spatter: Change the default data type to double

+++ spatter/scatter_gather_kernels.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-typedef uint64_t TElement;
+typedef double TElement;
 typedef uint64_t TIndex;
 
 void gather(TElement* __restrict__ dst, TElement* __restrict__ src, const TIndex* __restrict__ indices, const size_t array_size) {
